Rename low end module for evaluating derivatives
Since the low end module (`deriv`) share the name with the higher ned module
`eval_deriv`, it was renamed
