+++ gbasis/eval_deriv.py
@@ -1,7 +1,7 @@
 """Functions for evaluating Gaussian primitives."""
+from gbasis._deriv import _eval_deriv_contractions
 from gbasis.base_one import BaseOneIndex
 from gbasis.contractions import ContractedCartesianGaussians
-from gbasis.deriv import _eval_deriv_contractions
 import numpy as np
