+++ tests/test_deriv.py
@@ -1,7 +1,7 @@
-"""Test gbasis.deriv."""
+"""Test gbasis._deriv."""
 import itertools as it
 
-from gbasis.deriv import _eval_deriv_contractions
+from gbasis._deriv import _eval_deriv_contractions
 import numpy as np
 from utils import partial_deriv_finite_diff
 
@@ -31,7 +31,7 @@ def eval_deriv_prim(coord, orders, center, angmom_comps, alpha):
     Note
     ----
     If you want to evaluate the derivative of the contractions of the primitive, then use
-    `gbasis.deriv.eval_deriv_contraction` instead.
+    `gbasis._deriv.eval_deriv_contraction` instead.
 
     """
     return _eval_deriv_contractions(
@@ -67,16 +67,16 @@ def eval_prim(coord, center, angmom_comps, alpha):
     Note
     ----
     If you want to evaluate the contractions of the primitive, then use
-    `gbasis.deriv.eval_contraction` instead.
+    `gbasis._deriv.eval_contraction` instead.
 
     """
     return eval_deriv_prim(coord, np.zeros(angmom_comps.shape), center, angmom_comps, alpha)
 
 
 def test_eval_prim():
-    """Test gbasis.deriv.eval_prim.
+    """Test gbasis._deriv.eval_prim.
 
-    Note that this also tests the no derivative case of gbasis.deriv.eval_deriv_prim.
+    Note that this also tests the no derivative case of gbasis._deriv.eval_deriv_prim.
     """
     # angular momentum: 0
     assert np.allclose(
@@ -106,7 +106,7 @@ def test_eval_prim():
 
 
 def test_eval_deriv_prim():
-    """Test gbasis.deriv.eval_deriv_prim."""
+    """Test gbasis._deriv.eval_deriv_prim."""
     # first order
     for k in range(3):
         orders = np.zeros(3, dtype=int)
@@ -143,7 +143,7 @@ def test_eval_deriv_prim():
 
 
 def test_eval_contractions():
-    """Test gbasis.deriv._eval_deriv_contraction without derivatization."""
+    """Test gbasis._deriv._eval_deriv_contraction without derivatization."""
     # angular momentum: 0
     assert np.allclose(
         _eval_deriv_contractions(
@@ -258,7 +258,7 @@ def test_eval_contractions():
 
 
 def test_eval_deriv_contractions():
-    """Test gbasis.deriv._eval_deriv_contractions."""
+    """Test gbasis._deriv._eval_deriv_contractions."""
     # first order
     for k in range(3):
         orders = np.zeros(3, dtype=int)
@@ -403,7 +403,7 @@ def test_eval_deriv_contractions():
 
 
 def test_eval_deriv_generalized_contraction():
-    """Test gbasis.deriv._eval_deriv_contractions for generalized contractions."""
+    """Test gbasis._deriv._eval_deriv_contractions for generalized contractions."""
     for k, l in it.product(range(3), range(3)):
         orders = np.zeros(3, dtype=int)
         orders[k] += 1
